Stage 5: Extract Table component into smaller components
- Broke down monolithic 584-line Table.svelte into 5 focused components
- Created StandingsTable.svelte as main coordinator component
- Extracted ConferenceCard.svelte for conference-specific rendering
- Added TableSection.svelte for reusable table sections
- Created TableHeader.svelte for column headers with sorting
- Added TableRow.svelte for individual team rows
- Moved original component to archive with corrected imports
- Fixed linting issues with unused parameters and imports
- Updated ESLint config to allow console.error for proper error logging
- All 76 tests still pass and TypeScript compiles successfully
