+++ front-end/svelte/international-nhl-standings/src/components/Table/index.ts
@@ -1,0 +1,5 @@
+export { default as StandingsTable } from './StandingsTable.svelte';
+export { default as ConferenceCard } from './ConferenceCard.svelte';
+export { default as TableSection } from './TableSection.svelte';
+export { default as TableHeader } from './TableHeader.svelte';
+export { default as TableRow } from './TableRow.svelte';
